refactor: Move package for class management
